+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Types of changes
 )
 
 ## [3.3.2](https://github.com/pdir/mobilede-bundle/tree/3.3.2) – 2022-09-20
-- 
+- [Changed] change default value of field pdirVehicleFilterByAccount
+- [Fixed] fix assets dir (use web or public)
 
 ## [3.3.1](https://github.com/pdir/mobilede-bundle/tree/3.3.1) – 2022-08-02